+++ src/main/java/com/github/Cubolink/finalreality/model/character/player/CharacterClass/Black_Mage.java
@@ -5,7 +5,7 @@ import java.util.Random;
 import com.github.Cubolink.finalreality.model.character.ICharacter;
 import com.github.Cubolink.finalreality.model.statuseffects.Burned;
 import com.github.Cubolink.finalreality.model.statuseffects.Paralyzed;
-import com.github.Cubolink.finalreality.model.weapon.GenericWeapon;
+import com.github.Cubolink.finalreality.model.weapon.IWeapon;
 
 
 /**
@@ -19,15 +19,16 @@ public class Black_Mage extends AbstractCharacterClass implements IMage{
         super(classname);
     }
 
-    public Black_Mage(String classname, GenericWeapon weapon) {
+    public Black_Mage(String classname, IWeapon weapon) {
         super(classname, weapon);
     }
 
     /**
      * {@inheritDoc}
+     * @param weapon
      */
     @Override
-    public void equip(GenericWeapon weapon) {
+    public void equip(IWeapon weapon) {
         if (weapon.isWearableByMage()){
             this.equippedWeapon = weapon;
         }