+++ src/main/java/com/github/Cubolink/finalreality/model/weapon/AbstractWeapon.java
@@ -1,0 +1,70 @@
+package com.github.Cubolink.finalreality.model.weapon;
+
+import com.github.Cubolink.finalreality.model.AbstractItem;
+import com.github.Cubolink.finalreality.model.IWearableItem;
+
+/**
+ * An abstract class with common things that weapon have.
+ */
+public abstract class AbstractWeapon extends AbstractItem implements IWeapon, IWearableItem {
+    private final int physical_damage;
+
+    /**
+     * Instantiate a generic weapon.
+     * @param name its name.
+     * @param physical_damage its base damage when attacking physically.
+     * @param weight its weight.
+     */
+    public AbstractWeapon(String name, int physical_damage, double weight){
+        super(name, weight);
+        this.physical_damage = physical_damage;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public int getPhysicalDamage() {
+        return physical_damage;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public int getMagicalDamage(){
+        return 0;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean isWearableByEngineer() {
+        return false;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean isWearableByKnight(){
+        return false;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean isWearableByThief() {
+        return false;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean isWearableByMage() {
+        return false;
+    }
+}