+++ src/test/java/com/github/Cubolink/finalreality/model/character/player/CharacterClass/Black_MageTest.java
@@ -6,10 +6,7 @@ import com.github.Cubolink.finalreality.model.weapon.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.LinkedBlockingDeque;
 import java.util.concurrent.LinkedBlockingQueue;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -19,13 +16,14 @@ class Black_MageTest{
     //private List<ICharacter> testCharacters;
     private Black_Mage blackMageTest;
     private ICharacter opponent;
+    private IWeapon staff;
     protected BlockingQueue<ICharacter> turns;
 
     @BeforeEach
     void setUp(){
         turns = new LinkedBlockingQueue<>();
 
-        GenericWeapon staff = new Staff("Báculo", 5, 20, 6);
+        staff = new Staff("Báculo", 5, 20, 6);
         blackMageTest = new Black_Mage("Kuro mado-shi", staff);
         opponent = new Enemy(turns, "Enemigo", 40, 0, 10, 5);
     }
@@ -34,12 +32,11 @@ class Black_MageTest{
     void equip() {
         blackMageTest = new Black_Mage("Kuro mado-shi");
 
-        GenericWeapon axe = new Axe("Hacha", 20, 15);
-        GenericWeapon staff = new Staff("Báculo", 5, 20, 6);
-        GenericWeapon bow = new Bow("Arco de Hierro", 15, 6);
-        GenericWeapon knife = new Knife("Cuchillito", 10, 3);
-        GenericWeapon sword = new Sword("Espada", 15, 10);
-        GenericWeapon genericWeapon = new GenericWeapon("Piedra", 5, 1);
+        IWeapon axe = new Axe("Hacha", 20, 15);
+        IWeapon staff = new Staff("Báculo", 5, 20, 6);
+        IWeapon bow = new Bow("Arco de Hierro", 15, 6);
+        IWeapon knife = new Knife("Cuchillito", 10, 3);
+        IWeapon sword = new Sword("Espada", 15, 10);
 
         blackMageTest.equip(axe);
         assertNull(blackMageTest.getEquippedWeapon());
@@ -51,9 +48,6 @@ class Black_MageTest{
         assertNull(blackMageTest.getEquippedWeapon());
 
         blackMageTest.equip(sword);
-        assertNull(blackMageTest.getEquippedWeapon());
-
-        blackMageTest.equip(genericWeapon);
         assertNull(blackMageTest.getEquippedWeapon());
 
         blackMageTest.equip(staff);
@@ -77,7 +71,7 @@ class Black_MageTest{
     @Test
     void testEquals() {
         Staff other_staff = new Staff("Vara", 10, 5, 5);
-        Black_Mage other_blackMage = new Black_Mage("Mago negro avanzado");
+        Black_Mage other_blackMage = new Black_Mage("Mago negro avanzado", staff);
         Black_Mage same_class_name_same_weapon = new Black_Mage(
                 blackMageTest.getClassname(),blackMageTest.getEquippedWeapon());
         Black_Mage same_class_name_diff_weapon = new Black_Mage(
@@ -94,7 +88,7 @@ class Black_MageTest{
     @Test
     void testHashCode() {
         Staff other_staff = new Staff("Vara", 10, 5, 5);
-        Black_Mage other_blackMage = new Black_Mage("Mago negro avanzado");
+        Black_Mage other_blackMage = new Black_Mage("Mago negro avanzado", staff);
         Black_Mage same_class_name_same_weapon = new Black_Mage(
                 blackMageTest.getClassname(),blackMageTest.getEquippedWeapon());
         Black_Mage same_class_name_diff_weapon = new Black_Mage(