+++ src/test/java/com/github/Cubolink/finalreality/model/character/player/CharacterClass/EngineerTest.java
@@ -1,6 +1,5 @@
 package com.github.Cubolink.finalreality.model.character.player.CharacterClass;
 
-import com.github.Cubolink.finalreality.model.character.ICharacter;
 import com.github.Cubolink.finalreality.model.weapon.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -9,8 +8,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 class EngineerTest {
     private Engineer engineerTest;
-    private GenericWeapon axe;
-    private ICharacter opponent;
+    private IWeapon axe;
 
     @BeforeEach
     void setUp(){
@@ -22,20 +20,15 @@ class EngineerTest {
     void equip() {
         engineerTest = new Engineer("Kuro mado-shi");
 
-        GenericWeapon axe = new Axe("Hacha", 20, 15);
-        GenericWeapon staff = new Staff("Báculo", 5, 20, 6);
-        GenericWeapon bow = new Bow("Arco de Hierro", 15, 6);
-        GenericWeapon knife = new Knife("Cuchillito", 10, 3);
-        GenericWeapon sword = new Sword("Espada", 15, 10);
-        GenericWeapon genericWeapon = new GenericWeapon("Piedra", 5, 1);
+        IWeapon staff = new Staff("Báculo", 5, 20, 6);
+        IWeapon bow = new Bow("Arco de Hierro", 15, 6);
+        IWeapon knife = new Knife("Cuchillito", 10, 3);
+        IWeapon sword = new Sword("Espada", 15, 10);
 
         engineerTest.equip(knife);
         assertNull(engineerTest.getEquippedWeapon());
 
         engineerTest.equip(sword);
-        assertNull(engineerTest.getEquippedWeapon());
-
-        engineerTest.equip(genericWeapon);
         assertNull(engineerTest.getEquippedWeapon());
 
         engineerTest.equip(staff);
@@ -76,7 +69,7 @@ class EngineerTest {
         ICharacterClass other_character_class = new Knight("Cabashero", engineerTest.getEquippedWeapon());
 
         assertEquals(engineerTest.hashCode(), engineerTest.hashCode());
-        assertNotEquals(engineerTest.hashCode(), other_axe.hashCode());
+        assertNotEquals(engineerTest.hashCode(), other_engineer.hashCode());
         assertEquals(engineerTest.hashCode(), same_class_name_same_weapon.hashCode());
         assertEquals(engineerTest.hashCode(), same_class_name_diff_weapon.hashCode());
         assertNotEquals(engineerTest.hashCode(), other_character_class.hashCode());