+++ src/test/java/com/github/Cubolink/finalreality/model/weapon/AxeTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 
 class AxeTest {
-    private GenericWeapon axeTest;
+    private AbstractWeapon axeTest;
 
     @BeforeEach
     void setUp(){