+++ src/test/java/com/github/Cubolink/finalreality/model/weapon/BowTest.java
@@ -17,12 +17,12 @@ class BowTest {
     void testEquals() {
         assertEquals(bowTest, bowTest);
 
-        var notBow = new GenericWeapon("piedra", 3, 1);
+        var sword = new Sword("Espada de madera", 3, 5);
         var ballesta = new Bow("Bayesta", 15, 10);
         var bow_altered = new Bow(bowTest.getName(), bowTest.getPhysicalDamage()+10, bowTest.getWeight());
         var samebow = new Bow(bowTest.getName(), bowTest.getPhysicalDamage(), bowTest.getWeight());
 
-        assertNotEquals(bowTest, notBow);
+        assertNotEquals(bowTest, sword);
         assertNotEquals(bowTest, ballesta);
         assertNotEquals(bowTest, bow_altered);
         assertEquals(bowTest, samebow);
@@ -34,12 +34,12 @@ class BowTest {
     void testHashCode() {
         assertEquals(bowTest.hashCode(), bowTest.hashCode());
 
-        var notBow = new GenericWeapon(bowTest.getName(), bowTest.getPhysicalDamage(), bowTest.getWeight());
+        var sword = new Sword("Espada de madera", 3, 5);
         var ballesta = new Bow("Bayesta", 15, 10);
         var bow_altered = new Bow(bowTest.getName(), bowTest.getPhysicalDamage()+10, bowTest.getWeight());
         var samebow = new Bow(bowTest.getName(), bowTest.getPhysicalDamage(), bowTest.getWeight());
 
-        assertNotEquals(bowTest.hashCode(), notBow.hashCode());
+        assertNotEquals(bowTest.hashCode(), sword.hashCode());
         assertNotEquals(bowTest.hashCode(), ballesta.hashCode());
         assertNotEquals(bowTest.hashCode(), bow_altered.hashCode());
         assertEquals(bowTest.hashCode(), samebow.hashCode());