+++ src/test/java/com/github/Cubolink/finalreality/model/weapon/KnifeTest.java
@@ -18,13 +18,13 @@ class KnifeTest {
 
         assertEquals(knifeTest, knifeTest);
 
-        var notKnife = new GenericWeapon(knifeTest.getName(), knifeTest.getPhysicalDamage(), knifeTest.getWeight());
+        var sword = new Sword("Espada de madera", 3, 5);
         var cooking_knife = new Knife("Cuchillo", 5, 2);
         var altered_knife = new Knife(knifeTest.getName(), knifeTest.getPhysicalDamage()+5, knifeTest.getWeight());
         var same_knife = new Knife(knifeTest.getName(), knifeTest.getPhysicalDamage(), knifeTest.getWeight());
         var heavy_knife = new Knife(knifeTest.getName(), knifeTest.getPhysicalDamage(), knifeTest.getWeight()+10);
 
-        assertNotEquals(knifeTest, notKnife);
+        assertNotEquals(knifeTest, sword);
         assertNotEquals(knifeTest, cooking_knife);
         assertNotEquals(knifeTest, altered_knife);
         assertNotEquals(knifeTest, heavy_knife);
@@ -36,12 +36,12 @@ class KnifeTest {
     void testHashCode() {
         assertEquals(knifeTest.hashCode(), knifeTest.hashCode());
 
-        var notKnife = new GenericWeapon("piedra", 3, 1);
+        var sword = new Sword("Espada de madera", 3, 5);
         var cooking_knife = new Knife("Cuchillo", 5, 2);
         var altered_knife = new Knife(knifeTest.getName(), knifeTest.getPhysicalDamage()+5, knifeTest.getWeight());
         var same_knife = new Knife(knifeTest.getName(), knifeTest.getPhysicalDamage(), knifeTest.getWeight());
 
-        assertNotEquals(knifeTest.hashCode(), notKnife.hashCode());
+        assertNotEquals(knifeTest.hashCode(), sword.hashCode());
         assertNotEquals(knifeTest.hashCode(), cooking_knife.hashCode());
         assertNotEquals(knifeTest.hashCode(), altered_knife.hashCode());
         assertEquals(knifeTest.hashCode(), same_knife.hashCode());