+++ src/test/java/com/github/Cubolink/finalreality/model/weapon/StaffTest.java
@@ -17,14 +17,14 @@ class StaffTest {
     void testEquals() {
         assertEquals(staffTest, staffTest);
 
-        var notStaff = new GenericWeapon("piedra", 3, 1);
+        var sword = new Sword("Espada de madera", 3, 5);
         var rod = new Staff("Baculo", 5, 20, 5);
         var altered_staff = new Staff(staffTest.getName(), staffTest.getPhysicalDamage(),
                             staffTest.getMagicalDamage()+3, staffTest.getWeight());
         var same_staff = new Staff(staffTest.getName(), staffTest.getPhysicalDamage(),
                 staffTest.getMagicalDamage(), staffTest.getWeight());
 
-        assertNotEquals(staffTest, notStaff);
+        assertNotEquals(staffTest, sword);
         assertNotEquals(staffTest, rod);
         assertNotEquals(staffTest, altered_staff);
         assertEquals(staffTest, same_staff);
@@ -34,14 +34,14 @@ class StaffTest {
     void testHashCode() {
         assertEquals(staffTest.hashCode(), staffTest.hashCode());
 
-        var notStaff = new GenericWeapon("piedra", 3, 1);
+        var sword = new Sword("Espada de madera", 3, 5);
         var rod = new Staff("Baculo", 5, 20, 5);
         var altered_staff = new Staff(staffTest.getName(), staffTest.getPhysicalDamage(),
                 staffTest.getMagicalDamage()+3, staffTest.getWeight());
         var same_staff = new Staff(staffTest.getName(), staffTest.getPhysicalDamage(),
                 staffTest.getMagicalDamage(), staffTest.getWeight());
 
-        assertNotEquals(staffTest.hashCode(), notStaff.hashCode());
+        assertNotEquals(staffTest.hashCode(), sword.hashCode());
         assertNotEquals(staffTest.hashCode(), rod.hashCode());
         assertNotEquals(staffTest.hashCode(), altered_staff.hashCode());
         assertEquals(staffTest.hashCode(), same_staff.hashCode());