+++ src/test/java/com/github/Cubolink/finalreality/model/weapon/SwordTest.java
@@ -17,12 +17,12 @@ class SwordTest {
     void testEquals() {
         assertEquals(swordTest, swordTest);
 
-        var notSword = new GenericWeapon(swordTest.getName(), swordTest.getPhysicalDamage(), swordTest.getWeight());
+        var bow = new Bow(swordTest.getName(), swordTest.getPhysicalDamage(), swordTest.getWeight());
         var sable = new Sword("Sable", 25, 15);
         var altered_sword = new Sword(swordTest.getName(), swordTest.getPhysicalDamage()-10, swordTest.getWeight());
         var same_sword = new Sword(swordTest.getName(), swordTest.getPhysicalDamage(), swordTest.getWeight());
 
-        assertNotEquals(swordTest, notSword);
+        assertNotEquals(swordTest, bow);
         assertNotEquals(swordTest, sable);
         assertNotEquals(swordTest, altered_sword);
         assertEquals(swordTest, same_sword);
@@ -33,12 +33,12 @@ class SwordTest {
     void testHashCode() {
         assertEquals(swordTest.hashCode(), swordTest.hashCode());
 
-        var notSword = new GenericWeapon("piedra", 3, 1);
+        var bow = new Bow("piedra", 3, 1);
         var sable = new Sword("Sable", 25, 15);
         var altered_sword = new Sword(swordTest.getName(), swordTest.getPhysicalDamage()-10, swordTest.getWeight());
         var same_sword = new Sword(swordTest.getName(), swordTest.getPhysicalDamage(), swordTest.getWeight());
 
-        assertNotEquals(swordTest.hashCode(), notSword.hashCode());
+        assertNotEquals(swordTest.hashCode(), bow.hashCode());
         assertNotEquals(swordTest.hashCode(), sable.hashCode());
         assertNotEquals(swordTest.hashCode(), altered_sword.hashCode());
         assertEquals(swordTest.hashCode(), same_sword.hashCode());